Separated calculateTotalCost function to export for use in 2 different components
